sprites

+++ src/Core/Config.java
@@ -4,8 +4,8 @@ public class Config
 {
     public static final Boolean DEBUGMODE = false;
     public static final String MAPDEFINITION_EMPTY = "______";
-    public static final int GAMEWINDOWWIDTH = 1440;
-    public static final int GAMEWINDOWHEIGTH = 900;
+    public static final double GAMEWINDOWWIDTH = 1440;
+    public static final double GAMEWINDOWHEIGTH = 900;
     public static final int CAMERA_WIDTH = 1200;
     public static final int CAMERA_HEIGTH = 800;
